delete productHero component
