+++ support-frontend/assets/helpers/subscriptionsForms/formReducer.js
@@ -33,7 +33,7 @@ function createFormReducer(
   const { productPrices, orderIsAGift } = window.guardian;
 
   const initialState = {
-    stage: 'checkout',
+    stage: 'thankyou',
     product,
     title: null,
     email: user.email || '',
