+++ support-frontend/assets/pages/digital-subscription-checkout/components/thankYou/hero.jsx
@@ -4,14 +4,29 @@
 
 import React from 'react';
 
-
-import ProductHero, {
-  type GridImages,
-  type ImagesByCountry,
-} from 'components/productHero/productHero';
+import GridPicture, {
+  type GridImage,
+  type GridSlot,
+  type Source as GridSource,
+} from 'components/gridPicture/gridPicture';
+import { type ImageId as GridId } from 'helpers/images/theGrid';
+import LeftMarginSection from 'components/leftMarginSection/leftMarginSection';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
 
 // ----- Setup ----- //
+
+type GridImages = {
+  breakpoints: {
+    mobile: GridImage,
+    tablet: GridImage,
+    desktop: GridImage,
+  },
+  fallback: GridId,
+};
+
+type ImagesByCountry = {
+  [CountryGroupId]: GridImages,
+};
 
 const defaultHeroes: GridImages = {
   breakpoints: {
@@ -65,15 +80,51 @@ const heroesByCountry: ImagesByCountry = {
   AUDCountries: australiaHeroes,
 };
 
-const ThankYouHero = ({ countryGroupId }: {countryGroupId: CountryGroupId}) => (
-  <ProductHero
-    countryGroupId={countryGroupId}
-    imagesByCountry={heroesByCountry}
-    altText="digital subscription"
-    fallbackImgType="png"
-  />
-);
+type GridSlots = {
+  mobile: GridSlot,
+  tablet: GridSlot,
+  desktop: GridSlot,
+};
 
+
+const gridSlots: GridSlots = {
+  mobile: {
+    sizes: '240px',
+    media: '(max-width: 739px)',
+  },
+  tablet: {
+    sizes: '407px',
+    media: '(min-width: 740px) and (max-width: 1139px)',
+  },
+  desktop: {
+    sizes: '809px',
+    media: '(min-width: 1140px)',
+  },
+};
+
+const ThankYouHero = ({ countryGroupId }: {countryGroupId: CountryGroupId}) => {
+  const gridImages: GridImages = heroesByCountry[countryGroupId];
+
+  const sources: GridSource[] = [
+    { ...gridSlots.mobile, ...gridImages.breakpoints.mobile },
+    { ...gridSlots.tablet, ...gridImages.breakpoints.tablet },
+    { ...gridSlots.desktop, ...gridImages.breakpoints.desktop },
+  ];
+
+  return (
+    <div className="component-product-hero">
+      <LeftMarginSection>
+        <GridPicture
+          sources={sources}
+          fallback={gridImages.fallback}
+          fallbackSize={500}
+          altText="Digital subscription"
+          fallbackImgType="png"
+        />
+      </LeftMarginSection>
+    </div>
+  );
+};
 
 // ----- Export ----- //
