+++ support-frontend/assets/pages/paper-subscription-checkout/_legacyImports.scss
@@ -1,5 +1,4 @@
 @import '~stylesheets/gu-sass/gu-sass';
 
-@import '~components/productHero/productHero';
 @import '~components/progressMessage/progressMessage';
 @import '~components/generalErrorMessage/generalErrorMessage';
